requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
icalendar==5.0.11
httpx==0.26.0
jinja2==3.1.3
python-multipart==0.0.6
pydantic==2.5.3
python-dateutil==2.8.2
